+++ urllib3/_collections.py
@@ -247,13 +247,15 @@ class HTTPHeaderDict(MutableMapping):
         for key, value in kwargs.items():
             self.add(key, value)
 
-    def getlist(self, key):
+    def getlist(self, key, default=__marker):
         """Returns a list of all the values for the named field. Returns an
         empty list if the key doesn't exist."""
         try:
             vals = self._container[key.lower()]
         except KeyError:
-            return []
+            if default is self.__marker:
+                return []
+            return default
         else:
             return vals[1:]
 
@@ -261,6 +263,9 @@ class HTTPHeaderDict(MutableMapping):
     getheaders = getlist
     getallmatchingheaders = getlist
     iget = getlist
+
+    # Backwards compatibility for http.cookiejar
+    get_all = getlist
 
     def __repr__(self):
         return "%s(%s)" % (type(self).__name__, dict(self.itermerged()))